Simple regression provided

+++ EmSELab1/EmSELab1/DTOs/InputDataDto.cs
@@ -2,10 +2,10 @@
 {
     public class InputDataDto
     {
-        public int Y { get; set; }
+        public float Y { get; set; }
 
-        public int X1 { get; set; }
+        public float X1 { get; set; }
 
-        public int X2 { get; set; }
+        public float X2 { get; set; }
     }
 }
